fix:完善删除逻辑

+++ springboot/src/main/java/com/item/arrangement/controller/VideoController.java
@@ -48,8 +48,15 @@ public class VideoController extends BaseController{
     //删除视频
     @DeleteMapping("/{id}") //占位符 /{aa}/{bb} ==> (@PathVariable aa的类型 aa,@PathVariable bb的类型 bb)
     public Result<?> delete(@PathVariable Long id){
-        videoMapper.deleteById(id);
-        return Result.success();
+
+        Video res=videoMapper.selectById(id);
+        if(getAccount().getType().equals("user")&&res.getAuthorId()!=getAccount().getId()) return Result.error("401","无权限");
+        else {
+            videoMapper.deleteById(id);
+            FileController.removeOldAvatar(res.getCover());
+            FileController.removeOldAvatar(res.getSrc());
+            return Result.success();
+        }
     }
 
     //分页查询
